requirements.txt
pywinauto==0.6.8
pytest==7.4.2
python-dateutil